+++ Backend/ShoesCustom/ShoesPee/ShoesPee/src/main/java/com/ToDoiVar/ShoesPee/Services/UserService.java
@@ -6,9 +6,9 @@ import java.util.List;
 public interface UserService {
     public List<User> getAllUser();
     public User createUser(User user);
-    public User getUserById(Long id);
-    public void removeUser(Long id);
-    public User editUser(Long id,User newUser);
+    public User getUserById(int id);
+    public void removeUser(int id);
+    public User editUser(int id,User newUser);
     public User fineUserByName(String name);
 
 }
