+++ Backend/ShoesCustom/ShoesPee/ShoesPee/src/main/java/com/ToDoiVar/ShoesPee/repositiory/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User,Long>,CrudRepository<User,Long> {
+public interface UserRepository extends JpaRepository<User, Integer>,CrudRepository<User,Integer> {
 
     User findByUsername(String userName);
 }
